centralize autosuggest test code and make it flexible to aria-owns/aria-controls
