+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/component/SearchField.scala
@@ -2,7 +2,9 @@ package org.broadinstitute.dsde.firecloud.component
 
 import org.openqa.selenium.WebDriver
 
-case class SearchField(queryString: QueryString)(implicit webDriver: WebDriver) extends Component(queryString) {
+case class SearchField(queryString: QueryString)(implicit webDriver: WebDriver)
+  extends Component(queryString) with Suggests {
+
   def setText(text: String): Unit = {
     searchField(query).value = text
   }
