+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/page/workspaces/methodconfigs/WorkspaceMethodConfigDetailsPage.scala
@@ -139,23 +139,11 @@ class WorkspaceMethodConfigDetailsPage(namespace: String, name: String, methodCo
 
   def clickAndReadSuggestions(field: String): Seq[String] = {
     val dataTestId = s"$field-text-input"
-    val listOptionXpath = s"//div[@data-test-id='$field-suggestions']/ul[@role='listbox']/li[@role='option']"
-
     click on find(testId("test.hello.response-text-input")).get // force page scrolls down
 
-    val suggestionTextfield: Element = find(testId(dataTestId)).get
-    val expandedDropdownId = suggestionTextfield.underlying.getAttribute("aria-owns")
-
-    click on suggestionTextfield // click in textfield to invoke list dropdown
-    await condition find(testId(dataTestId)).exists(_.underlying.getAttribute("aria-expanded") == "true")
-
-    // wait for dropdown to contain at least one item
-    await condition {
-      find(xpath(s"//div[@id='$expandedDropdownId']")).exists(_.isDisplayed)
-      findAll(xpath(listOptionXpath)).map(_.text).toSeq.nonEmpty // getting Element's text force screen scroll if item is outside of viewport
-    }
-
-    findAll(xpath(listOptionXpath)).map(_.text).toSeq
+    val suggestionTextfield = TextField(TestId(dataTestId))
+    click on suggestionTextfield.query.element.underlying // activate the suggestions
+    suggestionTextfield.getSuggestions()
   }
 
   def readFieldValue(field: String): String = {
